Make tests discoverable by pytest
Wrap the various test_*.py contents in a `test_foo` function.
As it was, invoking pytest would import all the scripts, executing
the tests, but because they were's associated with a marked function
for class, it wasn't able to report any results.

Add a __main__ section invoking the test functions if the scripts
are executed directly to maintain the old functionality.

Note `test_struct.py` just contains diagnostic print functions
and doesn't run any tests itself.